+++ tests/resources/configs/iec_pas_61174_0_standard.py
@@ -1,4 +1,5 @@
 minimal_record_v1 = {
+    "$schema": "https://metadata-standards.data.bas.ac.uk/bas-metadata-generator-configuration-schemas/v2/iec-pas-61174-0-v1.json",
     "route_name": "minimal-test-route",
     "waypoints": [
         {"id": 1001, "revision": 0, "position": {"lat": 5.5, "lon": 50.55}},
@@ -8,6 +9,7 @@ minimal_record_v1 = {
 }
 
 complete_record_v1 = {
+    "$schema": "https://metadata-standards.data.bas.ac.uk/bas-metadata-generator-configuration-schemas/v2/iec-pas-61174-0-v1.json",
     "route_author": "Constance Watson",
     "route_name": "complete-test-route",
     "route_status": "complete",