+++ tests/resources/configs/iso19115_2_standard.py
@@ -20,12 +20,24 @@ base_complex_record_v2 = deepcopy(_base_complex_record_v2)  # type: dict
 complete_record_v2 = deepcopy(_complete_record_v2)  # type: dict
 
 minimal_record_v3 = deepcopy(_minimal_record_v3)  # type: dict
+minimal_record_v3[
+    "$schema"
+] = "https://metadata-standards.data.bas.ac.uk/bas-metadata-generator-configuration-schemas/v2/iso-19115-2-v3.json"
 
 base_simple_record_v3 = deepcopy(_base_simple_record_v3)  # type: dict
+base_simple_record_v3[
+    "$schema"
+] = "https://metadata-standards.data.bas.ac.uk/bas-metadata-generator-configuration-schemas/v2/iso-19115-2-v3.json"
 
 base_complex_record_v3 = deepcopy(_base_complex_record_v3)  # type: dict
+base_complex_record_v3[
+    "$schema"
+] = "https://metadata-standards.data.bas.ac.uk/bas-metadata-generator-configuration-schemas/v2/iso-19115-2-v3.json"
 
 complete_record_v3 = deepcopy(_complete_record_v3)  # type: dict
+complete_record_v3[
+    "$schema"
+] = "https://metadata-standards.data.bas.ac.uk/bas-metadata-generator-configuration-schemas/v2/iso-19115-2-v3.json"
 
 configs_safe_v2 = {
     "minimal_v2": minimal_record_v2,